+++ src/Kiota.Builder/Writers/Java/CodeMethodWriter.cs
@@ -602,7 +602,8 @@ public class CodeMethodWriter : BaseElementWriter<CodeMethod, JavaConventionServ
         else
             WriteSerializerBodyForInheritedModel(method, inherits, parentClass, writer);
 
-        if (parentClass.Methods.Any(x => x.AccessedProperty?.IsOfKind(CodePropertyKind.AdditionalData) ?? false))
+        if (parentClass.Methods.Any(x => x.AccessedProperty?.IsOfKind(CodePropertyKind.AdditionalData) ?? false) ||
+        parentClass.Properties.Any(x => x?.IsOfKind(CodePropertyKind.AdditionalData) ?? false))
             writer.WriteLine($"writer.writeAdditionalData(this.get{CodePropertyKind.AdditionalData.ToString().ToFirstCharacterUpperCase()}());");
     }
     private void WriteSerializerBodyForUnionModel(CodeClass parentClass, CodeMethod method, LanguageWriter writer)